feat: circuit board selectors
