+++ packages/ui-react/src/modules/Chips/constants.js
